Ref: Created app/src/components/market/sections/common_sections and /allowance folder and moved files inside
