+++ app/src/components/market/sections/market_create/steps/items/funding_and_fee_step.tsx
@@ -44,13 +44,13 @@ import { CurrencySelector } from '../../../../common/currency_selector'
 import { DisplayArbitrator } from '../../../../common/display_arbitrator'
 import { GridTransactionDetails } from '../../../../common/grid_transaction_details'
 import { MarketScale } from '../../../../common/market_scale'
-import { SetAllowance } from '../../../../common/set_allowance'
 import { TradingFeeSelector } from '../../../../common/trading_fee_selector'
 import { TransactionDetailsCard } from '../../../../common/transaction_details_card'
 import { TransactionDetailsLine } from '../../../../common/transaction_details_line'
 import { TransactionDetailsRow, ValueStates } from '../../../../common/transaction_details_row'
 import { VerifiedRow } from '../../../../common/verified_row'
 import { WarningMessage } from '../../../../common/warning_message'
+import { SetAllowance } from '../../../common_sections/allowance/set_allowance'
 import { Outcome } from '../outcomes'
 
 const CreateCardTop = styled(CreateCard)`
